src/main/java/Bedroom.java
import java.util.ArrayList;

public class Bedroom extends Room {

    private int roomNumber;
    private RoomType roomType;

    public Bedroom(ArrayList<Guest> guestList, int roomNumber, RoomType roomType ) {
        super(roomType.getCapacity(), guestList);
        this.roomNumber = roomNumber;
        this.roomType = roomType;
    }


    public int getRoomNumber() {
        return roomNumber;
    }

    public void setRoomNumber(int roomNumber) {
        this.roomNumber = roomNumber;
    }

    public RoomType getRoomType() {
        return roomType;
    }

    public void setRoomType(RoomType roomType) {
        this.roomType = roomType;
    }


}
